Move SetupSyncPointsToMockDirectIO to sync_point.h
